refactor: appとapplicationの統合
